feat: add data access

+++ src/core/Application/Zisrf.WorkoutDiary.Core.Application.Contracts/Dto/ActivityDto.cs
@@ -3,5 +3,5 @@
 public record ActivityDto(
     Guid Id,
     Guid ExerciseId,
-    int WorkingWeight,
+    double WorkingWeight,
     int RepetitionsCount);
